fan: Tweak the thresholds further
Dug through the upstream issue tracker, found other people complaining
about the fan noise problem.  It may just be luck of the draw, but I
would need to go find a replacement fan to test that which is annoying.

The person from the mfg side, just recommends to run the fan at 100%
above 55 C so that's what I'll do.